+++ Server/GameServer/PlayerInfos.h
@@ -152,21 +152,21 @@ public:
 	};
 	void		LevelUpPlayer()
 	{
-		this->pcProfile->avatarAttribute.wBasePhysicalOffence += this->byLevel_Up_Physical_Offence;
-		this->pcProfile->avatarAttribute.wBasePhysicalDefence += this->byLevel_Up_Physical_Defence;
-		this->pcProfile->avatarAttribute.wBaseEnergyOffence += this->byLevel_Up_Energy_Offence;
-		this->pcProfile->avatarAttribute.wBaseEnergyDefence += this->byLevel_Up_Energy_Defence;
-		this->pcProfile->avatarAttribute.byBaseCon += this->fLevel_Up_Con;
-		this->pcProfile->avatarAttribute.byBaseDex += this->fLevel_Up_Dex;
-		this->pcProfile->avatarAttribute.byBaseEng += this->fLevel_Up_Eng;
-		this->pcProfile->avatarAttribute.byBaseFoc += this->fLevel_Up_Foc;
-		this->pcProfile->avatarAttribute.byBaseSol += this->fLevel_Up_Sol;
-		this->pcProfile->avatarAttribute.byBaseStr += this->fLevel_Up_Str;
-		this->pcProfile->avatarAttribute.wBaseMaxRP += this->byLevel_Up_RP;
-		this->pcProfile->avatarAttribute.wBaseMaxEP = (this->pcProfile->avatarAttribute.wBaseMaxEP + this->byLevel_Up_EP);
-		this->pcProfile->avatarAttribute.wBaseMaxEP = this->pcProfile->avatarAttribute.wBaseMaxEP + (this->pcProfile->avatarAttribute.byLastEng * 4.2);
-		this->pcProfile->avatarAttribute.wBaseMaxLP = (this->pcProfile->avatarAttribute.wBaseMaxLP + this->byLevel_Up_LP);
-		this->pcProfile->avatarAttribute.wBaseMaxLP = this->pcProfile->avatarAttribute.wBaseMaxLP + (this->pcProfile->avatarAttribute.byLastCon* 4.7);
+		this->pcProfile->avatarAttribute.wBasePhysicalOffence += (WORD)this->byLevel_Up_Physical_Offence;
+		this->pcProfile->avatarAttribute.wBasePhysicalDefence += (WORD)this->byLevel_Up_Physical_Defence;
+		this->pcProfile->avatarAttribute.wBaseEnergyOffence += (WORD)this->byLevel_Up_Energy_Offence;
+		this->pcProfile->avatarAttribute.wBaseEnergyDefence += (WORD)this->byLevel_Up_Energy_Defence;
+		this->pcProfile->avatarAttribute.byBaseCon += (BYTE)this->fLevel_Up_Con;
+		this->pcProfile->avatarAttribute.byBaseDex += (BYTE)this->fLevel_Up_Dex;
+		this->pcProfile->avatarAttribute.byBaseEng += (BYTE)this->fLevel_Up_Eng;
+		this->pcProfile->avatarAttribute.byBaseFoc += (BYTE)this->fLevel_Up_Foc;
+		this->pcProfile->avatarAttribute.byBaseSol += (BYTE)this->fLevel_Up_Sol;
+		this->pcProfile->avatarAttribute.byBaseStr += (BYTE)this->fLevel_Up_Str;
+		this->pcProfile->avatarAttribute.wBaseMaxRP += (WORD)this->byLevel_Up_RP;
+		this->pcProfile->avatarAttribute.wBaseMaxEP = (WORD)(this->pcProfile->avatarAttribute.wBaseMaxEP + this->byLevel_Up_EP);
+		this->pcProfile->avatarAttribute.wBaseMaxEP = this->pcProfile->avatarAttribute.wBaseMaxEP + (WORD)(this->pcProfile->avatarAttribute.byLastEng * 4.2);
+		this->pcProfile->avatarAttribute.wBaseMaxLP = (this->pcProfile->avatarAttribute.wBaseMaxLP + (WORD)this->byLevel_Up_LP);
+		this->pcProfile->avatarAttribute.wBaseMaxLP = this->pcProfile->avatarAttribute.wBaseMaxLP + (WORD)(this->pcProfile->avatarAttribute.byLastCon* 4.7);
 		if (this->pcProfile->byLevel == 5)
 			this->CurRPBall = 1;
 		if (this->pcProfile->byLevel == 15)
